Add pagination (non interactive) on the  command
Legacy call of listtasks will now return the first 50 tasks. Optional parameters have been added: offset is the index of the first task to return, limit is the index of the last task (non inclusive) to return. Those two parameters must be used together. The tag filter is decoupled from the pagination.